all: configure health check support for all sc-bos apps
This makes it possible for autos, drivers, and zones to contribute health checks, and for the sc-bos server to expose those checks via the different APIs like HealthApi, HealthHistoryApi, and DevicesApi.

Systems get a special system.HealthCheckCollection which allows them to contribute (or remove) health checks to devices without going through the registry. Proxies and gateways will want to do this to avoid duplicate history and processing.