pkg/app: respond to smartcore.bos.{api}.v1.{service} requests
As part of our migration to a versioned API, this change introduces a renaming interceptor that translates requests like `/smartcore.bos.meter.v1.MeterApi/GetMeterReading` to the current format of `/smartcore.bos.MeterApi/GetMeterReading`. While we don't have any APIs that use this pattern yet, when we do they may need to live in a cohort with node that have yet to update. This change means an old node can serve requests from a new client, as long as the old node includes this change.