tools/genproto,go.mod,pkg/gen: update router and wrapper protoc plugins
This updated versions include a new usePaths=true options with disables the sc-golang specific fixed path output logic. This means we can start to utilise the semi-standard module and paths options for go, grpc, router, and wrapper to remove out custom fixings.

Updating has changed the filenames of the router and wrap.pb.go files, so this is included in this change.