policy: match trait gRPC service names exactly
In the policy (rego) files, we interpret any service starting with (string prefix) a trait name to be a trait API. This works fine for the current set of traits but is basically a coincidence, which won't be true once trait packages are reorganised.

The policies receive the list of trait names as data in the policy evaluation context, derived from the `alltraits` package registry. That registry actually contains all the  service names in the gRPC ServiceDesc which we should use instead.

1. Change the data injected into the policy evaluation context to contain gRPC service names associated with each trait, not just the trait names.
2. Perform exact matching against the known trait service names to determine if an incoming request is a trait request.